+++ NEWS.md
@@ -22,7 +22,8 @@
 * `confint_betabinom()`, `confint_fisher()`: Added argument `b_lives` which allows the user to specify probabilities of interest.
 * Added `print.rank_regression()`, `print.ml_estimation()`, `print.model_estimation()` and `print.model_estimation_list()`.
 * Added `vcov.model_estimation()`.
-* Added function `dist_delay()`: Generalizes the handling of delays. 
+* Added function `dist_delay()`: Generalizes the distribution-specific modeling of delays. 
+* Added function `mcs_delay()`: Generalizes the adjustment of operating times by delays.
 * Added lifecycle badges
 
 ## Documentation
@@ -34,6 +35,7 @@
 * `plot_layout()`
 * `mr_method()`, `johnson_method()`, `kaplan_method()` and `nelson_method()`: Use `estimate_cdf()` instead.
 * `dist_delay_register()` and `dist_delay_report()` are deprecated. Use `dist_delay()` instead. 
+* `mcs_delay_register()`, `mcs_delay_report()` and `mcs_delays()`  are deprecated. Use `mcs_delay()` instead.
 ## Minor improvements and bug fixes
 * Fixed bug inside `plot_mod_mix()` for the case of no mixture distribution
 * Fixed bug inside `confint_betabinom()`; many cases near one -> unique()